backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.12.0
chromadb==0.4.22
langchain==0.1.0
PyPDF2==3.0.1
beautifulsoup4==4.12.2
requests==2.31.0
numpy==1.24.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
alembic==1.12.1